MODORDSTOR-96 Add test class to suit
